backend/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-rest-auth==0.9.5
djangorestframework==3.9.0
gunicorn==20.0.4
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2-binary==2.8.5
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
drf-writable-nested==0.6.0
django-rest-swagger==2.2.0
